requirements.txt
tensorflow==2.13.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.29.0
requests==2.31.0

psycopg2-binary==2.9.9
joblib==1.3.2
folium
streamlit-folium